requirements.txt
aiogram>=3.4.0
aiosqlite>=0.20.0
python-dotenv>=1.0.0
pandas>=2.0.0
openpyxl>=3.0.0
fastapi>=0.110.0
uvicorn>=0.29.0
